+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/Common/NavigationScreenDataSource.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol NavigationScreenDataSource {
+    var destinationViewController: UIViewController { get }
+    var title: String { get }
+}