+++ changelog/mediation/mobileads-mytarget/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 5.17.5.1
+
+#### Updated
+* Updated minimum supported Yandex Mobile Ads SDK version to 5.9.0
+
 ## Version 5.17.5.0
 
 #### Updated